requirements.txt
fastapi==0.65.2
aiohttp==3.7.4.post0
gunicorn==20.1.0
Hypercorn==0.11.2
Pillow==8.3.1
python-multipart==0.0.5
requests==2.25.1
uvicorn==0.14.0
numpy==1.21.2
